refactor(platform): PDF系Stimulus Controllerをベースクラスに統合 (issue#232)
3つのPDF処理コントローラ(amex/bank/invoice)の共通ロジックを
base_pdf_upload_controllerに集約し、各コントローラは差分のみ定義。
